Declare Dependencies (Case 126792)
Some dependencies were missing altogether and two were only suggested, as they are only needed for optional features of the WfdMetaBundle. (In a perfect world, these optional features might be better suited in something like a bridge bundle.)

Nowadays I don't think the suggests is a good idea in our context:

If you want to use the optional features, they'll have constraints for their suggested/required packages. These constraints should be considered in the dependency graph of the project that uses the WfdMetaBundle. We cannot expect the project author to know these constraints and to declare them in the project's composer.json - it is possible they aren't direct dependencies of their project.

Declaring these constraints might have unexpected consequences when updating the WfdMetaBundle: new transivite package might be installed or their installation might fail if no set of working dependencies could be determined. But in practice, this problem seems to be manageable: doctrine/orm will be used in most of our projects anyway and the webfactory/http-cache-bundle has only dependencies you'll find in most of our projects, too.