Make the `always_expire_wfd_meta_resources` config setting work regardless of `kernel.debug` mode (#50)
The `always_expire_wfd_meta_resources` config setting can be used to
make `WfdMetaResource` instances always expire immediately. It is
supposed to be set in Symfony environments like `test` (e. g. in a
`config_test.yml` file or a `when@test` section).

However, until now, it works correctly only for `kernel.debug = true`. 

This is because it used a special `CacheBustingResourceChecker` that
would be used by the "inner" (regular) `ConfigCache` instance only. That
ResourceChecker would invalidate the cache when an instance of a
`WfdMetaResource` is found.

The `ConfigCache`, however, contains a shortcut to never validate the
cache in the first place when `kernel.debug` is disabled.

Thus, we need to move this behavior to the `WfdMetaConfigCache`, which
will run its check also in non-debug mode. There, we want to consider
the setting only in the case that there are actually `WfdMetaResource`s
associated with the cache in question.

Being able to disable `kernel.debug` mode for tests execution can
provide a significant speed-up in environments where you don't actually
modify code between test runs, i. e. a CI pipeline.